pyproject.toml
[tool.poetry]
name = "what_is_the_media_saying"
version = "0.1.0"
description = ""
authors = ["Matheus Westhelle <matheus.westhelle@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9"
Scrapy = "2.4.1"
elasticsearch = "7.11.0"
python-dateutil = "2.8.1"

[tool.poetry.dev-dependencies]
pytest = "5.2"
black = "20.8b1"
flake8 = "3.8.4"
pre-commit = "2.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
